+++ assemblyadapter-common-recycler-divider/src/androidTest/java/com/github/panpf/assemblyadapter/common/recycler/divider/test/internal/GridItemDividerProviderTest.kt
@@ -42,9 +42,7 @@ class GridItemDividerProviderTest {
                 .toItemDividerConfig(context),
             sideDividerConfig = DividerConfig.Builder(Divider.space(20)).build()
                 .toItemDividerConfig(context),
-            sideHeaderDividerConfig = DividerConfig.Builder(Divider.space(25)).build()
-                .toItemDividerConfig(context),
-            sideFooterDividerConfig = DividerConfig.Builder(Divider.space(30)).build()
+            sideHeaderAndFooterDividerConfig = DividerConfig.Builder(Divider.space(25)).build()
                 .toItemDividerConfig(context),
         )
 
@@ -110,7 +108,7 @@ class GridItemDividerProviderTest {
         position = 1
         Assert.assertEquals(0, getDivider(ItemDivider.Type.START)?.widthSize ?: 0)
         Assert.assertEquals(10, getDivider(ItemDivider.Type.TOP)?.widthSize ?: 0)
-        Assert.assertEquals(30, getDivider(ItemDivider.Type.END)?.widthSize ?: 0)
+        Assert.assertEquals(30, getDivider(ItemDivider.Type.END)?.widthSize ?: 0)// todo fix test error
         Assert.assertEquals(5, getDivider(ItemDivider.Type.BOTTOM)?.widthSize ?: 0)
 
         position = 2