+++ assemblyadapter-common-recycler-divider/src/main/java/com/github/panpf/assemblyadapter/recycler/divider/GridDividerItemDecoration.kt
@@ -122,19 +122,18 @@ open class GridDividerItemDecoration(
         private var useDividerAsFooterDivider = false
 
         private var sideDividerConfig: DividerConfig? = null
-        private var sideHeaderDividerConfig: DividerConfig? = null
-        private var sideFooterDividerConfig: DividerConfig? = null
-        private var useSideDividerAsSideHeaderDivider = false
-        private var useSideDividerAsSideFooterDivider = false
+        private var sideHeaderAndFooterDividerConfig: DividerConfig? = null
+        private var useSideDividerAsSideHeaderAndFooterDivider = false
 
         fun build(): GridDividerItemDecoration {
             return GridDividerItemDecoration(buildItemDividerProvider())
         }
 
         private fun buildItemDividerProvider(): GridItemDividerProvider {
-            if ((useSideDividerAsSideHeaderDivider || useSideDividerAsSideFooterDivider) && sideDividerConfig == null) {
+            if (useSideDividerAsSideHeaderAndFooterDivider && sideDividerConfig == null) {
                 throw IllegalArgumentException("Must call the sideDivider() method to configure the sideDivider")
             }
+            // todo Ensure that the size of sideDividerConfig is consistent with the size of sideHeaderAndFooterDividerConfig
 
             val finalDividerConfig =
                 dividerConfig ?: context.obtainStyledAttributes(
@@ -156,11 +155,8 @@ open class GridDividerItemDecoration(
                     ?: if (useDividerAsFooterDivider) finalDividerConfig else null)
                     ?.toItemDividerConfig(context),
                 sideDividerConfig = sideDividerConfig?.toItemDividerConfig(context),
-                sideHeaderDividerConfig = (sideHeaderDividerConfig
-                    ?: if (useSideDividerAsSideHeaderDivider) sideDividerConfig else null)
-                    ?.toItemDividerConfig(context),
-                sideFooterDividerConfig = (sideFooterDividerConfig
-                    ?: if (useSideDividerAsSideFooterDivider) sideDividerConfig else null)
+                sideHeaderAndFooterDividerConfig = (sideHeaderAndFooterDividerConfig
+                    ?: if (useSideDividerAsSideHeaderAndFooterDivider) sideDividerConfig else null)
                     ?.toItemDividerConfig(context),
             )
         }
@@ -315,22 +311,20 @@ open class GridDividerItemDecoration(
          * Set the header divider on the side of the item. You can configure to disable the divider or
          * provide a personalized divider in some cases through the [configBlock] function
          */
+        @Deprecated("Please use sideHeaderAndFooterDivider instead", ReplaceWith("sideHeaderAndFooterDivider"))
         fun sideHeaderDivider(
             divider: Divider,
             configBlock: (DividerConfig.Builder.() -> Unit)? = null
         ): Builder {
-            this.sideHeaderDividerConfig = DividerConfig.Builder(divider).apply {
-                configBlock?.invoke(this)
-            }.build()
-            return this
+            return sideHeaderAndFooterDivider(divider, configBlock)
         }
 
         /**
          * Set the header divider on the side of the item
          */
+        @Deprecated("Please use sideHeaderAndFooterDivider instead", ReplaceWith("sideHeaderAndFooterDivider"))
         fun sideHeaderDivider(config: DividerConfig): Builder {
-            this.sideHeaderDividerConfig = config
-            return this
+            return sideHeaderAndFooterDivider(config)
         }
 
 
@@ -338,22 +332,20 @@ open class GridDividerItemDecoration(
          * Set the footer divider on the side of the item. You can configure to disable the divider or
          * provide a personalized divider in some cases through the [configBlock] function
          */
+        @Deprecated("Please use sideHeaderAndFooterDivider instead", ReplaceWith("sideHeaderAndFooterDivider"))
         fun sideFooterDivider(
             divider: Divider,
             configBlock: (DividerConfig.Builder.() -> Unit)? = null
         ): Builder {
-            this.sideFooterDividerConfig = DividerConfig.Builder(divider).apply {
-                configBlock?.invoke(this)
-            }.build()
-            return this
+            return sideHeaderAndFooterDivider(divider, configBlock)
         }
 
         /**
          * Set the footer divider on the side of the item
          */
+        @Deprecated("Please use sideHeaderAndFooterDivider instead", ReplaceWith("sideHeaderAndFooterDivider"))
         fun sideFooterDivider(config: DividerConfig): Builder {
-            this.sideFooterDividerConfig = config
-            return this
+            return sideHeaderAndFooterDivider(config)
         }
 
 
@@ -365,10 +357,7 @@ open class GridDividerItemDecoration(
             divider: Divider,
             configBlock: (DividerConfig.Builder.() -> Unit)? = null
         ): Builder {
-            this.sideHeaderDividerConfig = DividerConfig.Builder(divider).apply {
-                configBlock?.invoke(this)
-            }.build()
-            this.sideFooterDividerConfig = DividerConfig.Builder(divider).apply {
+            this.sideHeaderAndFooterDividerConfig = DividerConfig.Builder(divider).apply {
                 configBlock?.invoke(this)
             }.build()
             return this
@@ -378,8 +367,7 @@ open class GridDividerItemDecoration(
          * Set the header and footer divider on the side of the item
          */
         fun sideHeaderAndFooterDivider(config: DividerConfig): Builder {
-            this.sideHeaderDividerConfig = config
-            this.sideFooterDividerConfig = config
+            this.sideHeaderAndFooterDividerConfig = config
             return this
         }
 
@@ -387,25 +375,24 @@ open class GridDividerItemDecoration(
         /**
          * Use side divider as the header side divider
          */
+        @Deprecated("Please use useSideDividerAsSideHeaderAndFooterDivider instead", ReplaceWith("useSideDividerAsSideHeaderAndFooterDivider"))
         fun useSideDividerAsSideHeaderDivider(use: Boolean = true): Builder {
-            this.useSideDividerAsSideHeaderDivider = use
-            return this
+            return useSideDividerAsSideHeaderAndFooterDivider(use)
         }
 
         /**
          * Use side divider as the footer side divider
          */
+        @Deprecated("Please use useSideDividerAsSideHeaderAndFooterDivider instead", ReplaceWith("useSideDividerAsSideHeaderAndFooterDivider"))
         fun useSideDividerAsSideFooterDivider(use: Boolean = true): Builder {
-            this.useSideDividerAsSideFooterDivider = use
-            return this
+            return useSideDividerAsSideHeaderAndFooterDivider(use)
         }
 
         /**
          * Use side divider as the header and footer side divider
          */
         fun useSideDividerAsSideHeaderAndFooterDivider(use: Boolean = true): Builder {
-            this.useSideDividerAsSideHeaderDivider = use
-            this.useSideDividerAsSideFooterDivider = use
+            this.useSideDividerAsSideHeaderAndFooterDivider = use
             return this
         }
     }